Implement different formatting options for message text
Implement three different options for formatting the text:
- using the shr library
- using markdown mode
- using the original gitter.el formatting 'mechanism'

Additionally add the command `gitter-cycle-formatting-library` for cycling
between the different options (requires re-visiting the messages buffer)

Markdown mode already slightly improves on the original gitter.el formatting,
but it only fontifies code (e.g. links are fontified but not working)

SHR probably does the best job. It does not yet format code blocks, but links
are working. For this case improving the formatting means implementing the
improvements in shr.el, which is nice as it means improvements to an Emacs core
library.